+++ .eslintrc.js
@@ -1,9 +1,9 @@
 const { join } = require("path");
-const ESLintPresets = require("@manuth/eslint-plugin-typescript");
+const { PluginName, PresetName } = require("@manuth/eslint-plugin-typescript");
 
 module.exports = {
     extends: [
-        `plugin:${ESLintPresets.PluginName}/${ESLintPresets.PresetName.RecommendedWithTypeChecking}`
+        `plugin:${PluginName}/${PresetName.RecommendedWithTypeChecking}`
     ],
     env: {
         node: true,
@@ -11,9 +11,9 @@ module.exports = {
     },
     parserOptions: {
         project: [
-            join(__dirname, "tsconfig.json"),
+            join(__dirname, "tsconfig.app.json"),
             join(__dirname, "tsconfig.eslint.json"),
-            join(__dirname, "tsconfig.webpack.json"),
+            join(__dirname, "tsconfig.web.json"),
             join(__dirname, "src", "test", "tsconfig.json"),
             join(__dirname, "src", "tests", "tsconfig.json")
         ]