+++ src/test/essentials.test.ts
@@ -1,8 +1,10 @@
 import { EssentialTests } from "../tests/essentials";
+import { CommonHooks } from "./CommonHooks";
 
 suite(
     "Essential Tests",
     () =>
     {
+        CommonHooks();
         EssentialTests();
     });