+++ CHANGELOG.md
@@ -6,7 +6,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## MarkdownConverter [Unreleased]
 
-[Show differences](https://github.com/manuth/MarkdownConverter/compare/v5.1.2...dev)
+[Show differences](https://github.com/manuth/MarkdownConverter/compare/v5.1.3...dev)
+
+## MarkdownConverter v5.1.3
+### Updated
+  - All dependencies
+  - Drone CI configuration for improving automated publishing
+  - Deprecated `vscode-test` package
+  - `tsconfig.json` files to improve development experience
+
+### Removed
+  - Dependabot workflows
+
+### Added
+  - Support for running unit-tests using the `Mocha Test Explorer`
+
+[Show differences](https://github.com/manuth/MarkdownConverter/compare/v5.1.2...v5.1.3)
 
 ## MarkdownConverter v5.1.2
 ### Fixed